Add example to retrieve file content in statistics API
The OpenAPI definition has been updated to correctly reflect the fact
that the statistics endpoint returns a octet-stream.
Since the schema is not "file", the generated client doesn't expect a
file but a 'str'. Hence we need to save the content of the response
manually to a file, and retrieve the filename in the response header.

Show how to retrieve the name of the downloaded file from the response
headers.