Take deleted files into consideration when generating new clients
Bug Description:
When files are removed in the new generated client, it would not be detected and will not be committed.
This is caused by copying and overriding new code on the cloned directories.

Fix:
Move the git directory from the cloned client to a new generated client,
so that we can detect all the changes including delete files and new files.